# src/value_stream/app/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "ModelSettings-Output": {
        "additionalProperties": false,
        "properties": {
          "deployment_cadence": {
            "default": 5,
            "minimum": 0.0,
            "title": "Deployment Cadence",
            "type": "integer"
          },
          "deployment_duration": {
            "default": 0.25,
            "minimum": 0.0,
            "title": "Deployment Duration",
            "type": "number"
          },
          "deployment_failure_rate": {
            "default": 0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Deployment Failure Rate",
            "type": "number"
          },
          "distribution": {
            "default": "linear",
            "enum": [
              "linear",
              "normal"
            ],
            "title": "Distribution",
            "type": "string"
          },
          "efficiency_max": {
            "default": 1.5,
            "exclusiveMinimum": 0.0,
            "title": "Efficiency Max",
            "type": "number"
          },
          "efficiency_min": {
            "default": 0.5,
            "exclusiveMinimum": 0.0,
            "title": "Efficiency Min",
            "type": "number"
          },
          "qa_failure_cost": {
            "default": 0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Qa Failure Cost",
            "type": "number"
          },
          "qa_failure_rate": {
            "default": 0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Qa Failure Rate",
            "type": "number"
          },
          "qa_size": {
            "default": 2,
            "exclusiveMinimum": 0.0,
            "title": "Qa Size",
            "type": "integer"
          },
          "qa_time_cost": {
            "default": 0.1,
            "minimum": 0.0,
            "title": "Qa Time Cost",
            "type": "number"
          },
          "support_interval": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Support Interval"
          },
          "support_task_story_points": {
            "default": 1,
            "minimum": 0.0,
            "title": "Support Task Story Points",
            "type": "number"
          },
          "team_size": {
            "default": 4,
            "exclusiveMinimum": 0.0,
            "title": "Team Size",
            "type": "integer"
          },
          "toolchain_size": {
            "default": 1,
            "exclusiveMinimum": 0.0,
            "title": "Toolchain Size",
            "type": "integer"
          }
        },
        "required": [
          "team_size",
          "efficiency_min",
          "efficiency_max",
          "distribution",
          "deployment_cadence",
          "qa_size",
          "qa_time_cost",
          "qa_failure_rate",
          "qa_failure_cost",
          "toolchain_size",
          "deployment_duration",
          "deployment_failure_rate",
          "support_interval",
          "support_task_story_points"
        ],
        "title": "ModelSettings",
        "type": "object"
      },
...
      "OutcomeSummary": {
        "additionalProperties": false,
        "properties": {
          "cached": {
            "default": false,
            "title": "Cached",
            "type": "boolean"
          },
          "completion_time": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completion Time"
          },
          "cursor": {
            "default": 0,
            "title": "Cursor",
            "type": "integer"
          },
          "delivered_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivered Value"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorEnvelope"
              },
              {
                "type": "null"
              }
            ]
          },
          "loss_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Loss Percent"
          },
          "scenario": {
            "$ref": "#/components/schemas/ConcreteScenario"
          },
          "status": {
            "default": "queued",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "scenario",
          "status",
          "cursor",
          "cached",
          "loss_percent",
          "completion_time",
          "delivered_value",
          "error"
        ],
        "title": "OutcomeSummary",
        "type": "object"
      },
...
      "RunRequest": {
        "additionalProperties": false,
        "properties": {
          "definition_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Definition Id"
          },
          "family_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Value"
          },
          "intent": {
            "default": "manual",
            "enum": [
              "manual",
              "interactive"
            ],
            "title": "Intent",
            "type": "string"
          },
          "name": {
            "default": "Baseline",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "preview_digest": {
            "title": "Preview Digest",
            "type": "string"
          },
          "property": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Property"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "source_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Run Id"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "request_id",
          "preview_digest"
        ],
        "title": "RunRequest",
        "type": "object"
      },
      "RunStatus": {
        "additionalProperties": false,
        "properties": {
          "cancel_requested": {
            "default": false,
            "title": "Cancel Requested",
            "type": "boolean"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorEnvelope"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "intent": {
            "title": "Intent",
            "type": "string"
          },
          "job_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "last_cursor": {
            "default": 0,
            "title": "Last Cursor",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "outcomes": {
            "items": {
              "$ref": "#/components/schemas/OutcomeSummary"
            },
            "title": "Outcomes",
            "type": "array"
          },
          "pinned": {
            "default": false,
            "title": "Pinned",
            "type": "boolean"
          },
          "retry_paused": {
            "default": false,
            "title": "Retry Paused",
            "type": "boolean"
          },
          "state": {
            "enum": [
              "submitting",
              "queued",
              "running",
              "reconnecting",
              "cancelling",
              "completed",
              "completed_with_errors",
              "cancelled",
              "failed"
            ],
            "title": "State",
            "type": "string"
          },
          "superseded": {
            "default": false,
            "title": "Superseded",
            "type": "boolean"
          },
          "task_set_id": {
            "format": "uuid",
            "title": "Task Set Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "task_set_id",
          "state",
          "intent",
          "job_id",
          "outcomes",
          "last_cursor",
          "error",
          "pinned",
          "superseded",
          "cancel_requested",
          "retry_paused"
        ],
        "title": "RunStatus",
        "type": "object"
      },
...
      "ScenarioDefinition-Output": {
        "additionalProperties": false,
        "properties": {
          "execution_seed": {
            "default": 456,
            "maximum": 9007199254740991.0,
            "minimum": 0.0,
            "title": "Execution Seed",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "default": "Delivery model",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "revision": {
            "default": 1,
            "exclusiveMinimum": 0.0,
            "title": "Revision",
            "type": "integer"
          },
          "settings": {
            "$ref": "#/components/schemas/ModelSettings-Output"
          },
          "sweeps": {
            "additionalProperties": {
              "anyOf": [
                {
                  "items": {
                    "anyOf": [
                      {
                        "type": "number"
                      },
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "type": "array"
                },
                {
                  "$ref": "#/components/schemas/SweepRange"
                }
              ]
            },
            "title": "Sweeps",
            "type": "object"
          },
          "team_seed": {
            "default": 123,
            "maximum": 9007199254740991.0,
            "minimum": 0.0,
            "title": "Team Seed",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "revision",
          "settings",
          "sweeps",
          "team_seed",
          "execution_seed"
        ],
        "title": "ScenarioDefinition",
        "type": "object"
      },
...
      "TaskSet": {
        "additionalProperties": false,
        "properties": {
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "generator_version": {
            "default": "1",
            "title": "Generator Version",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "revision": {
            "default": 1,
            "title": "Revision",
            "type": "integer"
          },
          "spec": {
            "$ref": "#/components/schemas/TaskSetSpec-Output"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskInput"
            },
            "title": "Tasks",
            "type": "array"
          }
        },
        "required": [
          "id",
          "revision",
          "spec",
          "content_hash",
          "generator_version",
          "tasks"
        ],
        "title": "TaskSet",
        "type": "object"
      },
...
      "TaskSetSpec-Output": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "default": 100,
            "exclusiveMinimum": 0.0,
            "title": "Count",
            "type": "integer"
          },
          "depreciation_rate": {
            "default": 0.005,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Depreciation Rate",
            "type": "number"
          },
          "initial_value": {
            "discriminator": {
              "mapping": {
                "constant": "#/components/schemas/ConstantValue-Output",
                "uniform": "#/components/schemas/UniformValue-Output"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/ConstantValue-Output"
              },
              {
                "$ref": "#/components/schemas/UniformValue-Output"
              }
            ],
            "title": "Initial Value"
          },
          "name": {
            "default": "Task set",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "seed": {
            "default": 42,
            "maximum": 9007199254740991.0,
            "minimum": 0.0,
            "title": "Seed",
            "type": "integer"
          },
          "story_points": {
            "discriminator": {
              "mapping": {
                "constant": "#/components/schemas/ConstantValue-Output",
                "uniform": "#/components/schemas/UniformValue-Output"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/ConstantValue-Output"
              },
              {
                "$ref": "#/components/schemas/UniformValue-Output"
              }
            ],
            "title": "Story Points"
          }
        },
        "required": [
          "name",
          "count",
          "story_points",
          "initial_value",
          "depreciation_rate",
          "seed"
        ],
        "title": "TaskSetSpec",
        "type": "object"
      },
...
      "Workspace": {
        "additionalProperties": false,
        "properties": {
          "baseline_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Id"
          },
          "current_task_set": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Task Set"
          },
          "definitions": {
            "items": {
              "$ref": "#/components/schemas/ScenarioDefinition-Output"
            },
            "title": "Definitions",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "latest_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Id"
          },
          "revision": {
            "default": 0,
            "title": "Revision",
            "type": "integer"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunStatus"
            },
            "title": "Runs",
            "type": "array"
          },
          "task_sets": {
            "items": {
              "$ref": "#/components/schemas/TaskSet"
            },
            "title": "Task Sets",
            "type": "array"
          },
          "task_spec": {
            "$ref": "#/components/schemas/TaskSetSpec-Output"
          }
        },
        "required": [
          "id",
          "revision",
          "task_spec",
          "definitions",
          "task_sets",
          "current_task_set",
          "baseline_id",
          "latest_id",
          "runs"
        ],
        "title": "Workspace",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Value Stream Application",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/workspaces": {
      "post": {
        "operationId": "create_workspace_api_v1_workspaces_post",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Create Workspace"
      }
    },
...
    "/api/v1/workspaces/{workspace_id}/runs": {
      "post": {
        "operationId": "start_run_api_v1_workspaces__workspace_id__runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Start Run"
      }
    },
    "/api/v1/workspaces/{workspace_id}/runs/{run_id}": {
      "delete": {
        "operationId": "cancel_api_v1_workspaces__workspace_id__runs__run_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Cancel"
      },
      "get": {
        "operationId": "run_api_v1_workspaces__workspace_id__runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Run"
      }
    },
...
  }
}